% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunce_cells.R
\name{analyze_outcome_immune}
\alias{analyze_outcome_immune}
\title{Title: Immunce cells as exposure}
\usage{
analyze_outcome_immune(outcome_id, max_retries = 50)
}
\arguments{
\item{max_retries}{}
}
\value{
result
}
\description{
Title: Immunce cells as exposure
}
\examples{
for (outcome_id in outcome_ids){
write.table(analyze_outcome_immune(outcome_id=outcome_id,max_retries = 50),file = paste0(outcome_id,".txt"),sep = "\t",quote = F,row.names = F)
}
}
