% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_instruments.R
\name{extract_instruments}
\alias{extract_instruments}
\title{Extract Instruments for Given Outcomes}
\usage{
extract_instruments(
  outcomes,
  p1 = 5e-08,
  clump = TRUE,
  p2 = 5e-08,
  r2 = 0.001,
  kb = 10000,
  access_token = ieugwasr::check_access_token(),
  force_server = FALSE
)
}
\arguments{
\item{outcomes}{A character vector of outcomes.}

\item{p1}{A numeric threshold for p-value.}

\item{clump}{A logical indicating whether to perform clumping.}

\item{p2}{A secondary p-value threshold for clumping.}

\item{r2}{A numeric threshold for the r-squared value.}

\item{kb}{A numeric indicating the window size in kilobases for clumping.}

\item{access_token}{A token to access the `ieugwasr` database. By default, it checks for a valid access token.}

\item{force_server}{A logical indicating whether to force the server connection.}
}
\value{
A data frame with extracted instrument variables.
}
\description{
This function extracts instruments based on given outcomes using data from `ieugwasr` package.
}
\examples{
\dontrun{
  result <- extract_instruments(outcomes = c("BMI", "LDL"), p1 = 5e-8)
}
}
