% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data for Mendelian Randomization analysis}
\usage{
format_data(
  dat,
  type = "exposure",
  snps = NULL,
  header = TRUE,
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  z_col = "z",
  info_col = "info",
  chr_col = "chr",
  pos_col = "pos",
  log_pval = FALSE
)
}
\arguments{
\item{dat}{The input dataset.}

\item{type}{Character, defaulting to "exposure". Specifies the type of the data.}

\item{snps}{Vector, default NULL. If not NULL, filters the data to include only specified SNPs.}

\item{header}{Logical, indicating if the data contains a header.
... [Other parameters like phenotype_col, snp_col, etc. Each should have a brief description.]}

\item{log_pval}{Logical. If TRUE, calculates p-values from the log transformed values.}
}
\value{
A formatted dataset suitable for MR analysis.
}
\description{
This function harmonizes and adjusts the input dataset for MR analysis by renaming, coercing,
and filtering columns according to the specified parameters.
}
\examples{
# [Include a simple usage example here]

}
\seealso{
[Any related functions or external resources]
}
\author{
[Author's Name]
}
