% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effect.R
\name{get_effect}
\alias{get_effect}
\title{get_effect function}
\usage{
get_effect(directory, immune_ref)
}
\arguments{
\item{directory}{A character string specifying the directory path that
contains the .rda files.}

\item{immune_ref}{A data frame that is used for merging with the final effect size
estimates data frame.}
}
\value{
A data frame containing the effect size estimates for all .rda files.
}
\description{
This function processes .rda files from a specified directory and generates
effect size estimates. For each .rda file, a subdirectory is created if it
does not already exist. The effect size estimates are then calculated for each
subdirectory and written to a .csv file. The function returns the effect size
estimates for all .rda files as a data frame.
}
\examples{
get_effect("~/path_to_directory", immune_ref_data)

}
