% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_finn.R
\name{get_finn}
\alias{get_finn}
\title{Harmonise and process FinnGen R9 data.}
\usage{
get_finn(finn_dir = "finn", save_dir = "finn_r", cores = 30)
}
\arguments{
\item{finn_dir}{The directory where the FinnGen data files are located.}

\item{save_dir}{Directory where processed data should be saved. Default is "finn_r".}

\item{cores}{The number of cores to use for parallel computation. Default is 2.}
}
\description{
This function processes and harmonises the FinnGen R9 data. The function reads each file,
formats and harmonises it, and then saves the processed data in a specified directory.
It uses parallel computation to process the list of files.
}
\examples{
\dontrun{
get_finn(finn_dir = "finn", save_dir = "finn_r", cores = 30)
}
}
