% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_immc.R
\name{get_immc}
\alias{get_immc}
\title{Retrieve Data for 731 Immune Cells}
\usage{
get_immc(
  use_preprocessed = TRUE,
  p1 = 1e-05,
  p2 = 5e-08,
  count_try_max = 50,
  r2 = 0.001,
  kb = 10000,
  mc_cores = 30
)
}
\arguments{
\item{use_preprocessed}{Logical. If TRUE, the function will return the preprocessed data.
If FALSE, the function will process the data using the provided parameters.
Default is TRUE.}

\item{p1}{Threshold parameter. Default is 1e-5.}

\item{p2}{Second threshold parameter. Default is 5e-08.}

\item{count_try_max}{Maximum number of tries. Default is 50.}

\item{r2}{R squared threshold parameter. Default is 0.001.}

\item{kb}{Distance threshold parameter in kilobases. Default is 10000.}

\item{mc_cores}{Number of cores to use for parallel processing. Default is 30.}
}
\value{
A data frame with extracted data for the 731 immune cells.
}
\description{
This function retrieves data for 731 immune cells. It can either return
preprocessed data directly or reprocess the data using provided parameters.
}
\examples{
\dontrun{
# Using preprocessed data
immc_data_preprocessed <- get_immc()

# Processing data with custom parameters
immc_data_custom <- get_immc(use_preprocessed = FALSE, p1 = 5e-5, mc_cores = 15)
}
}
