% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local.R
\name{get_local}
\alias{get_local}
\title{Convert local VCF file to Two-Sample MR input format and export it as a CSV file}
\usage{
get_local(vcf_file, type)
}
\arguments{
\item{vcf_file}{The local VCF file to be converted}

\item{type}{The type of GWAS study, "exposure" or "outcome"}
}
\value{
Export the converted data as a CSV file
}
\description{
Convert local VCF file to Two-Sample MR input format and export it as a CSV file
}
\examples{
get_local("ieu-a-2.vcf.gz", "exposure")
}
