% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunce_cells.R
\name{immune_total_check}
\alias{immune_total_check}
\title{Title immune_total_check}
\usage{
immune_total_check(exposure_id, outcome_name, outcome_source)
}
\arguments{
\item{outcome_source}{}
}
\description{
Title immune_total_check
}
