% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsmr.R
\name{mr_Speed}
\alias{mr_Speed}
\title{Perform Speedy Mendelian Randomization Analysis}
\usage{
mr_Speed(dat = harmonised_dat, prop_var_explained = T, sample = NULL)
}
\arguments{
\item{dat}{A data frame containing the necessary variables for the MR analysis.
It should at least have columns for SNP, exposure effect size, exposure standard error,
outcome effect size, and outcome standard error. Default is `harmonised_dat`.}

\item{prop_var_explained}{A logical value indicating whether to compute the proportion
of variance explained by the instrumental variables. Default is TRUE.}

\item{sample}{An optional sample data or criteria to subset the main `dat` data frame
for analysis. Useful for sensitivity analyses or when working with a subset of the data.
Default is NULL, which means the whole dataset will be used.}
}
\value{
mr_res A list or data frame containing the results of the MR analysis.
This will include estimates of causal effects, confidence intervals, and other relevant
statistical measures. (Note: Update this description based on the actual structure and
content of the return value.)
}
\description{
This function conducts a rapid Mendelian Randomization (MR) analysis on the provided dataset.
MR is a method of using genetic variations to determine causal relationships between exposures and outcomes.
}
\examples{
\dontrun{
harmonised_data <- data.frame( ... ) # example data structure
result <- mr_Speed(harmonised_data, prop_var_explained = TRUE, sample = NULL)
}
}
