% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsmr.R
\name{p.adjust}
\alias{p.adjust}
\title{Adjust p-values for Multiple Comparisons}
\usage{
p.adjust(p, method = p.adjust.methods, n = length(p))
}
\arguments{
\item{p}{A numeric vector of raw p-values that you want to adjust.}

\item{method}{A character string indicating the adjustment method to be used.
Possible values include all methods available in `p.adjust.methods`.
The default is to use all methods provided by `p.adjust.methods`.}

\item{n}{An integer specifying the number of comparisons,
typically the number of p-values or the length of the `p` vector. Default is the length of `p`.}
}
\value{
adjusted_p A numeric vector of adjusted p-values. (Note: Update this description
based on the actual structure and content of the return value.)
}
\description{
This function adjusts p-values for multiple comparisons using a specified method.
Adjusting for multiple comparisons can control the family-wise error rate or the false discovery rate.
}
\examples{
\dontrun{
raw_p_values <- c(0.01, 0.05, 0.1)
adjusted_values <- p.adjust(raw_p_values, method = "holm")
}
}
