% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunce_cells.R
\name{analyze_outcome_immune}
\alias{analyze_outcome_immune}
\title{Title: Immunce cells as exposure}
\usage{
analyze_outcome_immune(outcome_id, max_retries = 50)
}
\arguments{
\item{outcome_id}{A unique identifier for the outcome}

\item{max_retries}{Maximum number of retries for the analysis}
}
\value{
result The result of the analysis
}
\description{
Title: Immunce cells as exposure
}
\examples{
\dontrun{
analyze_outcome_immune(outcome_id = "some_id", max_retries = 50)
}
}
