% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunce_cells.R
\name{immune_total_check}
\alias{immune_total_check}
\title{Title immune_total_check}
\usage{
immune_total_check(exposure_id, outcome_name, outcome_source)
}
\arguments{
\item{exposure_id}{A unique identifier for the exposure}

\item{outcome_name}{The name of the outcome}

\item{outcome_source}{The source of the outcome}
}
\description{
Title immune_total_check
}
\examples{
\dontrun{
library(dplyr)
library(TwoSampleMR)
#https://gwas.mrcieu.ac.uk/datasets/
outcome_name = "finn-b-N14_ENDOMETRIOSIS"
outcome_source = "finn"
iddf=read.table(paste0(outcome_name,".txt"),header = T,sep = "\t")
# 暴露数据的SNP，工具变量
exposure_id=iddf$id.exposure
immune_total_check(exposure_id=exposure_id,
outcome_name=outcome_name,
outcome_source=outcome_source)
}
}
